+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ## [Unreleased]
 
+## [0.2.0] - 2021-09-25
+
+- Modify UI
+- Add slider bar for image width & height
+- Add side bar for more detail adjustment
+  - image tile grid gap
+  - image display mode
+    - cover (css)
+    - contain (css)
+    - fill (css)
+  - show label checkbox (close #19)
+- Change slider range
+- Fix to display appropriate panel icon
+
 ## [0.1.3] - 2021-02-02
 
 - Fix critical bug about displaying and opening image.